feat(versionable): refactor to more easily accomodate additional versions
