+++ test/configValidation.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const tape = require('tape');
-
 const configValidation = require('../configValidation');
 
 module.exports.tests = {};
