+++ test/version/2.4/document.js
@@ -1,4 +1,4 @@
-var schema = require('../mappings/document');
+var schema = require('../../../version/2.4/mappings/document');
 
 module.exports.tests = {};
