+++ version/2.4/mappings/partial/boundingbox.json
