+++ version/2.4/mappings/partial/centroid.js
