+++ version/2.4/mappings/partial/hash.json
