+++ version/2.4/mappings/partial/literal.json
