+++ version/2.4/mappings/partial/shape.json
