+++ version/2.4/settings.js
@@ -4,21 +4,21 @@ var fs = require('fs');
 var path = require('path');
 var merge = require('lodash.merge');
 var peliasConfig = require('pelias-config');
-var punctuation = require('./punctuation');
-var synonymFile = require('./synonyms/parser');
+var punctuation = require('../../punctuation');
+var synonymFile = require('../../synonyms/parser');
 
 // load synonyms from disk
-var synonyms = fs.readdirSync(path.join(__dirname, 'synonyms'))
+var synonyms = fs.readdirSync(path.join(__dirname, '../../synonyms'))
                  .sort()
                  .filter( f => f.match(/\.txt$/) )
                  .reduce(( acc, cur ) => {
                    acc[cur.replace('.txt','')] = synonymFile(
-                     path.join(__dirname, 'synonyms', cur)
+                     path.join(__dirname, '../../synonyms', cur)
                    );
                    return acc;
                  }, {});
 
-require('./configValidation').validate(peliasConfig.generate());
+require('../../configValidation').validate(peliasConfig.generate());
 
 function generate(){
   var config = peliasConfig.generate();
